+++ src/include/date.h
@@ -60,7 +60,7 @@ public:
 	bool ContainsDate(const CDate &date) const;	/// whether this date "contains" another (i.e. if it is subsequent to another, and in an appropriate timeline)
 	CDate ToBaseCalendar(const CCalendar *current_calendar) const;
 	std::string ToDisplayString(const CCalendar *calendar) const;
-	unsigned long GetTotalHours(const CCalendar *calendar) const;	/// gets the total amount of hours for the particular calendar in this date, counting from -10,000 in the base calendar
+	unsigned long long GetTotalHours(const CCalendar *calendar) const;	/// gets the total amount of hours for the particular calendar in this date, counting from -10,000 in the base calendar
 	
 	int Year;
 	char Month;
